requirements.txt
streamlit>=1.32.0
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.26.0
matplotlib==3.8.0
streamlit-authenticator==0.2.3
